- fixed the sync issue. - added the FSM/PA impl version. - fixed the initialisation value for offset.
