+++ src/main/scala/com/shah/persistence/query/model/QueryViewFSMImpl.scala
@@ -1,0 +1,10 @@
+package com.shah.persistence.query.model
+
+import akka.persistence.fsm.PersistentFSM
+import akka.persistence.fsm.PersistentFSM.FSMState
+
+trait QueryViewFSMImpl[S <: FSMState, D, E]
+  extends PersistentFSM[S,D,E] with QueryViewImplBase[D]{
+
+  override def applyEvent(domainEvent: E, currentData: D): D = ???
+}
